requirements.txt
pyproj
numpy
pyshp
dbfread
matplotlib==3.2
numba
dwdweather2


shapefile
patch






mpl_toolkits
pickle
json

